% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listWrappers.R
\name{listWrappers}
\alias{listWrappers}
\title{List all wrapper functions in SuperLearner}
\usage{
listWrappers(what = "both")
}
\arguments{
\item{what}{What list to return. Can be \code{both} for both prediction
algorithms and screening algorithms, \code{SL} for the prediction
algorithms, \code{screen} for the screening algorithms, \code{method} for
the estimation method details, or anything else will return a list of all
(exported) functions in the \code{SuperLearner} package. Additional wrapper
functions are available at
\url{https://github.com/ecpolley/SuperLearnerExtra}.}
}
\value{
Invisible character vector with all exported functions in the
\pkg{SuperLearner} package
}
\description{
List all wrapper functions in SuperLearner
}
\examples{

listWrappers(what = "SL")
listWrappers(what = "screen")
}
\seealso{
\code{\link{SuperLearner}}
}
\author{
Eric C Polley \email{epolley@uchicago.edu}
}
\keyword{utilities}
