% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcSuperLearner.R
\name{mcSuperLearner}
\alias{mcSuperLearner}
\alias{snowSuperLearner}
\title{SuperLearner with multicore support}
\usage{
mcSuperLearner(
  Y,
  X,
  newX = NULL,
  family = gaussian(),
  SL.library,
  method = "method.NNLS",
  id = NULL,
  verbose = FALSE,
  control = list(),
  cvControl = list(),
  obsWeights = NULL,
  env = parent.frame(),
  mc.cores = getOption("mc.cores", 2L)
)

snowSuperLearner(
  cluster,
  Y,
  X,
  newX = NULL,
  family = gaussian(),
  SL.library,
  method = "method.NNLS",
  id = NULL,
  verbose = FALSE,
  control = list(),
  cvControl = list(),
  obsWeights = NULL,
  env = parent.frame()
)
}
\arguments{
\item{Y, X, newX, family, SL.library, method, id, verbose, control, cvControl, obsWeights, env}{See \code{\link[=SuperLearner]{SuperLearner()}}.}

\item{mc.cores}{The number of cores to use, i.e. at most how many child processes will be run simultaneously. Passed to \code{parallel::mclapply()}.}

\item{cluster}{a \code{cluster} object from \code{parallel::makeCluster()}.}
}
\description{
\code{mcSuperlearner()} and \code{snowSuperlearner()} allow for multicore parallelization for \code{\link[=SuperLearner]{SuperLearner()}}.
}
\details{
\code{mcSuperlearner()} uses \code{parallel::mclapply()}, and \code{snowSuperlearner()} uses \code{parallel::parLapply()}.
}
