% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.bartMachine.R
\name{predict.SL.bartMachine}
\alias{predict.SL.bartMachine}
\title{bartMachine prediction}
\usage{
\method{predict}{SL.bartMachine}(object, newdata, family, X = NULL,
  Y = NULL, ...)
}
\description{
bartMachine prediction
}

