% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.biglasso.R
\name{predict.SL.biglasso}
\alias{predict.SL.biglasso}
\title{Prediction wrapper for SL.biglasso}
\usage{
\method{predict}{SL.biglasso}(object, newdata, ...)
}
\arguments{
\item{object}{SL.kernlab object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{Unused additional arguments}
}
\description{
Prediction wrapper for SL.biglasso objects.
}
\seealso{
\code{\link{SL.biglasso}} \code{\link[biglasso]{biglasso}}
  \code{\link[biglasso]{predict.biglasso}}
}
