% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.dbarts.R
\name{predict.SL.dbarts}
\alias{predict.SL.dbarts}
\title{dbarts prediction}
\usage{
\method{predict}{SL.dbarts}(object, newdata, family, ...)
}
\arguments{
\item{object}{SuperLearner object}

\item{newdata}{Dataframe to predict the outcome}

\item{family}{"gaussian" for regression, "binomial" for binary
classification. (Not used)}

\item{...}{Additional arguments (not used)}
}
\description{
WARNING: dbarts does not currently support predict(). Must use newX when
training via SL.dbarts.
}
