% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.extraTrees.R
\name{predict.SL.extraTrees}
\alias{predict.SL.extraTrees}
\title{extraTrees prediction on new data}
\usage{
\method{predict}{SL.extraTrees}(object, newdata, family, ...)
}
\arguments{
\item{object}{Model fit object from SuperLearner}

\item{newdata}{Dataframe}

\item{family}{Binomial or gaussian}

\item{...}{Any remaining arguments (not used).}
}
\description{
extraTrees prediction on new data
}
