% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.glm.R
\name{predict.SL.glm}
\alias{predict.SL.glm}
\title{Prediction for SL.glm}
\usage{
\method{predict}{SL.glm}(object, newdata, ...)
}
\arguments{
\item{object}{SL.glm object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{Unused additional arguments}
}
\description{
Prediction for SL.glm
}
\seealso{
\code{\link{SL.glm}} \code{\link[stats]{glm}}
  \code{\link[stats]{predict.glm}}  \code{\link{SL.speedglm}}
}
