% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.glmnet.R
\name{predict.SL.glmnet}
\alias{predict.SL.glmnet}
\title{Prediction for an SL.glmnet object}
\usage{
\method{predict}{SL.glmnet}(object, newdata, remove_extra_cols = T,
  add_missing_cols = T, ...)
}
\arguments{
\item{object}{Result object from SL.glmnet}

\item{newdata}{Dataframe or matrix that will generate predictions.}

\item{remove_extra_cols}{Remove any extra columns in the new data that were
not part of the original model.}

\item{add_missing_cols}{Add any columns from original data that do not exist
in the new data, and set values to 0.}

\item{...}{Any additional arguments (not used).}
}
\description{
Prediction for the glmnet wrapper.
}
\seealso{
\code{\link{SL.glmnet}}
}
