% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.kernelKnn.R
\name{predict.SL.kernelKnn}
\alias{predict.SL.kernelKnn}
\title{Prediction for SL.kernelKnn}
\usage{
\method{predict}{SL.kernelKnn}(object, newdata, ...)
}
\arguments{
\item{object}{SL.kernelKnn object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{Unused additional arguments}
}
\description{
Prediction for SL.kernelKnn
}
