% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.lm.R
\name{predict.SL.lm}
\alias{predict.SL.lm}
\title{Prediction for SL.lm}
\usage{
\method{predict}{SL.lm}(object, newdata, ...)
}
\arguments{
\item{object}{SL.lm object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{Unused additional arguments}
}
\description{
Prediction for SL.lm
}
\seealso{
\code{\link{SL.lm}} \code{\link[stats]{lm}}
  \code{\link[stats]{predict.lm}}  \code{\link{SL.speedlm}}
}
