% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.ranger.R
\name{predict.SL.ranger}
\alias{predict.SL.ranger}
\title{Prediction wrapper for ranger random forests}
\usage{
\method{predict}{SL.ranger}(object, newdata, family, num.threads = 1,
  verbose = object$verbose, ...)
}
\arguments{
\item{object}{SL.kernlab object}

\item{newdata}{Dataframe to generate predictions}

\item{family}{Gaussian or binomial}

\item{num.threads}{Number of threads used for parallelization}

\item{verbose}{If TRUE output additional information during execution.}

\item{...}{Unused additional arguments}
}
\description{
Prediction wrapper for SL.ranger objects.
}
\seealso{
\code{\link{SL.ranger}} \code{\link[ranger]{ranger}}
  \code{\link[ranger]{predict.ranger}}
}
