% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.speedglm.R
\name{predict.SL.speedglm}
\alias{predict.SL.speedglm}
\title{Prediction for SL.speedglm}
\usage{
\method{predict}{SL.speedglm}(object, newdata, ...)
}
\arguments{
\item{object}{SL.speedglm object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{Unused additional arguments}
}
\description{
Prediction for SL.speedglm
}
\seealso{
\code{\link{SL.speedglm}} \code{\link[speedglm]{speedglm}}
  \code{\link[speedglm]{predict.speedglm}}
}
