% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.speedlm.R
\name{predict.SL.speedlm}
\alias{predict.SL.speedlm}
\title{Prediction for SL.speedlm}
\usage{
\method{predict}{SL.speedlm}(object, newdata, ...)
}
\arguments{
\item{object}{SL.speedlm object}

\item{newdata}{Dataframe to generate predictions}

\item{...}{Unused additional arguments}
}
\description{
Prediction for SL.speedlm, a fast lm()
}
\seealso{
\code{\link{SL.speedlm}} \code{\link[speedglm]{speedlm}}
  \code{\link[speedglm]{predict.speedlm}}  \code{\link{SL.speedglm}}
}
