% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.SL.template.R
\name{predict.SL.template}
\alias{predict.SL.template}
\alias{write.predict.SL.template}
\title{Wrapper functions for computing predictions from SL learners}
\usage{
\method{predict}{SL.template}(object, newdata, family, X = NULL, Y = NULL, ...)

write.predict.SL.template(file = "", ...)
}
\arguments{
\item{object}{an object resulting from the fitting function.}

\item{newdata}{a data frame of predictors for which to compute predictions.}

\item{family}{Either \code{\link[=gaussian]{gaussian()}} or \code{\link[=binomial]{binomial()}} to describe the error distribution. Link function information will be ignored.}

\item{X}{The predictor variables in the training data set, usually a data.frame.}

\item{Y}{The outcome in the training data set. Must be a numeric vector.}

\item{\dots}{For \code{predict.SL} wrappers, other remaining arguments. For \code{write.predict.SL.template()}, arguments passed to \code{\link[=cat]{cat()}}.}

\item{file}{A connection, or a character string naming a file to print to.
Passed to \code{\link[=cat]{cat()}}.}
}
\value{
\code{predict.SL} wrappers return a vector of predictions for each unit in \code{newdata}.
}
\description{
Wrapper functions for computing predictions from SL learners
}
\examples{
write.predict.SL.template(file = '')
}
\keyword{utilities}
