% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.xgboost.R
\name{predict.SL.xgboost}
\alias{predict.SL.xgboost}
\title{XGBoost prediction on new data}
\usage{
\method{predict}{SL.xgboost}(object, newdata, family, ...)
}
\arguments{
\item{object}{Model fit object from SuperLearner}

\item{newdata}{Dataframe that will be converted to an xgb.DMatrix}

\item{family}{Binomial or gaussian}

\item{...}{Any remaining arguments (not supported though).}
}
\description{
XGBoost prediction on new data
}
