% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screen.R
\name{screen}
\alias{screen}
\alias{write.screen.template}
\alias{screen.template}
\alias{All}
\alias{screen.randomForest}
\alias{screen.SIS}
\alias{screen.ttest}
\alias{screen.corP}
\alias{screen.corRank}
\alias{screen.glmnet}
\title{Screening algorithms for SuperLearner}
\usage{
write.screen.template(file = "", ...)
}
\arguments{
\item{file}{A connection, or a character string naming a file to print to.
Passed to \code{\link{cat}}.}

\item{\dots}{Additional arguments passed to \code{\link{cat}}}
}
\value{
A logical vector with the length equal to the
number of columns in \code{X}.  TRUE indicates the variable (column of X)
should be included.
}
\description{
Screening algorithms for \code{SuperLearner} to be used with
\code{SL.library}.
}
\examples{
write.screen.template(file = '')

}
\seealso{
\code{\link{SuperLearner}}
}
\author{
Eric C Polley \email{polley.eric@mayo.edu}
}
\keyword{utilities}
