% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.bayesglm.R
\name{SL.bayesglm}
\alias{SL.bayesglm}
\alias{predict.SL.bayesglm}
\title{Wrapper for Bayesian GLM learner using \code{arm}}
\usage{
SL.bayesglm(Y, X, newX = X, family = gaussian(), obsWeights = NULL, ...)

\method{predict}{SL.bayesglm}(object, newdata, ...)
}
\arguments{
\item{Y}{The outcome in the training data set. Must be a numeric vector.}

\item{X}{The predictor variables in the training data set, usually a data.frame.}

\item{newX}{The predictor variables in the validation data set. The
structure should match X.}

\item{family}{Either \code{\link[=gaussian]{gaussian()}} or \code{\link[=binomial]{binomial()}} to
describe the error distribution. Link function information will be ignored.}

\item{obsWeights}{Optional observation weights.}

\item{...}{For SL wrappers, other remaining arguments. For \code{write.SL.template()}, arguments passed to \code{\link[=cat]{cat()}}.}

\item{object}{an object resulting from the fitting function.}

\item{newdata}{a data frame of predictors for which to compute predictions.}
}
\description{
Support Bayesian GLM via the \pkg{arm} package.
}
