% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.caret.R
\name{SL.caret}
\alias{SL.caret}
\title{SL wrapper for machine learning learner using \code{caret}}
\usage{
SL.caret(
  Y,
  X,
  newX = X,
  family = gaussian(),
  obsWeights = NULL,
  method = "rf",
  tuneLength = 3,
  trControl = caret::trainControl(method = "cv", number = 10, verboseIter = TRUE),
  metric = ifelse(family$family == "gaussian", "RMSE", "Accuracy"),
  ...
)
}
\arguments{
\item{Y}{The outcome in the training data set. Must be a numeric vector.}

\item{X}{The predictor variables in the training data set, usually a data.frame.}

\item{newX}{The predictor variables in the validation data set. The
structure should match X.}

\item{family}{Either \code{\link[=gaussian]{gaussian()}} or \code{\link[=binomial]{binomial()}} to
describe the error distribution. Link function information will be ignored.}

\item{obsWeights}{Optional observation weights.}

\item{method, tuneLength, trControl, metric}{arguments passed to \code{caret::train()}}

\item{...}{For SL wrappers, other remaining arguments. For \code{write.SL.template()}, arguments passed to \code{\link[=cat]{cat()}}.}
}
\description{
Support machine learning via the \pkg{caret} package.
}
\examples{
# how to change to a different method:
SL.caret.rpart <- function(..., method = "rpart") {
  SL.caret(..., method = method)
}
}
