% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.cforest.R
\name{SL.cforest}
\alias{SL.cforest}
\title{TODO: document all parameters.}
\usage{
SL.cforest(Y, X, newX, family, ntree = 1000, mtry = max(floor(ncol(X)/3),
  1), mincriterion = 0, teststat = "quad", testtype = "Univ",
  replace = F, fraction = 0.632, ...)
}
\description{
TODO: document all parameters.
}

