% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.earth.R
\name{SL.earth}
\alias{SL.earth}
\alias{predict.SL.earth}
\title{SL wrapper for \code{earth::earth()}}
\usage{
SL.earth(
  Y,
  X,
  newX = X,
  family = gaussian(),
  degree = 2,
  penalty = 3,
  nk = max(21, 2 * ncol(X) + 1),
  pmethod = "backward",
  nfold = 0,
  ncross = 1,
  minspan = 0,
  endspan = 0,
  ...
)

\method{predict}{SL.earth}(object, newdata, ...)
}
\arguments{
\item{Y}{The outcome in the training data set. Must be a numeric vector.}

\item{X}{The predictor variables in the training data set, usually a data.frame.}

\item{newX}{The predictor variables in the validation data set. The
structure should match X.}

\item{family}{Either \code{\link[=gaussian]{gaussian()}} or \code{\link[=binomial]{binomial()}} to
describe the error distribution. Link function information will be ignored.}

\item{degree, penalty, nk, pmethod, nfold, ncross, minspan, endspan}{Arguments passed to \code{earth::earth()}.}

\item{...}{For SL wrappers, other remaining arguments. For \code{write.SL.template()}, arguments passed to \code{\link[=cat]{cat()}}.}

\item{object}{an object resulting from the fitting function.}

\item{newdata}{a data frame of predictors for which to compute predictions.}
}
\description{
SL wrapper for \code{earth::earth()}
}
