% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.gam.R
\name{SL.gam}
\alias{SL.gam}
\alias{predict.SL.gam}
\title{SL wrapper for GAMs}
\usage{
SL.gam(
  Y,
  X,
  newX = X,
  family = gaussian(),
  obsWeights = NULL,
  deg.gam = 2,
  cts.num = 4,
  ...
)

\method{predict}{SL.gam}(object, newdata, ...)
}
\arguments{
\item{Y}{The outcome in the training data set. Must be a numeric vector.}

\item{X}{The predictor variables in the training data set, usually a data.frame.}

\item{newX}{The predictor variables in the validation data set. The
structure should match X.}

\item{family}{Either \code{\link[=gaussian]{gaussian()}} or \code{\link[=binomial]{binomial()}} to
describe the error distribution. Link function information will be ignored.}

\item{obsWeights}{Optional observation weights.}

\item{deg.gam}{Degree of smoothing (passed to the \code{df} argument of \code{gam::s()}).}

\item{cts.num}{Number of levels required for a variable to be consider continuous (and therefore given a smoothing spline).}

\item{...}{For SL wrappers, other remaining arguments. For \code{write.SL.template()}, arguments passed to \code{\link[=cat]{cat()}}.}

\item{object}{an object resulting from the fitting function.}

\item{newdata}{a data frame of predictors for which to compute predictions.}
}
\description{
Wrapper for generalized additivie models (GAMs) using \code{gam::gam()}.
}
\examples{
# easy to add additional algorithms with different degrees
}
