% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.lm.R
\name{SL.lm}
\alias{SL.lm}
\title{Wrapper for lm}
\usage{
SL.lm(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{Outcome variable}

\item{X}{Training dataframe}

\item{newX}{Test dataframe}

\item{family}{Gaussian or binomial}

\item{obsWeights}{Observation-level weights}

\item{...}{Any remaining arguments, not used.}
}
\description{
Wrapper for OLS via lm(), which may be faster than glm().
}
\seealso{
\code{\link{predict.SL.lm}} \code{\link[stats]{lm}}
  \code{\link[stats]{predict.lm}}  \code{\link{SL.speedlm}}
}
