% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.speedlm.R
\name{SL.speedlm}
\alias{SL.speedlm}
\title{Wrapper for speedlm}
\usage{
SL.speedlm(Y, X, newX, family, obsWeights, ...)
}
\arguments{
\item{Y}{Outcome variable}

\item{X}{Training dataframe}

\item{newX}{Test dataframe}

\item{family}{Gaussian or binomial}

\item{obsWeights}{Observation-level weights}

\item{...}{Any remaining arguments, not used.}
}
\description{
Speedlm is a fast version of lm()
}
\references{
Enea, M. A. R. C. O. (2013). Fitting linear models and generalized linear
models with large data sets in R. Statistical Methods for the Analysis of
Large Datasets: book of short papers, 411-414.
}
\seealso{
\code{\link{predict.SL.speedlm}} \code{\link[speedglm]{speedlm}}
  \code{\link[speedglm]{predict.speedlm}}  \code{\link{SL.speedglm}}
}
