% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.template.R
\name{SL.template}
\alias{SL.template}
\alias{SL.caret.rpart}
\alias{SL.gbm}
\alias{SL.ipredbagg}
\alias{SL.knn}
\alias{SL.leekasso}
\alias{SL.loess}
\alias{SL.logreg}
\alias{SL.nnet}
\alias{SL.nnls}
\alias{SL.polymars}
\alias{SL.randomForest}
\alias{SL.ridge}
\alias{SL.rpart}
\alias{SL.rpartPrune}
\alias{SL.step}
\alias{SL.step.forward}
\alias{SL.step.interaction}
\alias{SL.stepAIC}
\alias{SL.svm}
\alias{write.SL.template}
\title{Wrapper functions for prediction algorithms in SuperLearner}
\usage{
SL.template(Y, X, newX, family, obsWeights, id, ...)

write.SL.template(file = "", ...)
}
\arguments{
\item{Y}{The outcome in the training data set. Must be a numeric vector.}

\item{X}{The predictor variables in the training data set, usually a data.frame.}

\item{newX}{The predictor variables in the validation data set. The
structure should match X.}

\item{family}{Either \code{\link[=gaussian]{gaussian()}} or \code{\link[=binomial]{binomial()}} to
describe the error distribution. Link function information will be ignored.}

\item{obsWeights}{Optional observation weights.}

\item{id}{Optional cluster identification variable.}

\item{\dots}{For SL wrappers, other remaining arguments. For \code{write.SL.template()}, arguments passed to \code{\link[=cat]{cat()}}.}

\item{file}{A connection, or a character string naming a file to print to.
Passed to \code{\link[=cat]{cat()}}.}
}
\value{
SL wrappers return a list with two elements:
\item{pred}{ The predicted values for the
rows in \code{newX}. }
\item{fit}{ A list. Contains all objects necessary to
get predictions for new observations from specific algorithm. }
}
\description{
Template function for SuperLearner prediction wrappers and built in options.
}
\examples{
write.SL.template(file = '')

}
\seealso{
\code{\link{SuperLearner}}
}
\author{
Eric C Polley \email{epolley@uchicago.edu}
}
\keyword{utilities}
