% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearner-package.R
\docType{package}
\name{SuperLearner-package}
\alias{SuperLearner-package}
\title{SuperLearner: Super Learner Prediction}
\description{
Implements the super learner prediction method and contains a library of prediction algorithms to be used in the super learner.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ecpolley/SuperLearner}
  \item Report bugs at \url{https://github.com/ecpolley/SuperLearner/issues}
}

}
\author{
\strong{Maintainer}: Eric Polley \email{epolley@uchicago.edu}

Authors:
\itemize{
  \item Erin LeDell
  \item Chris Kennedy
  \item Mark van der Laan [thesis advisor]
}

Other contributors:
\itemize{
  \item Sam Lendle [contributor]
}

}
\keyword{internal}
