% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SuperLearnerNews.R
\name{SuperLearnerDocs}
\alias{SuperLearnerDocs}
\alias{SuperLearnerNews}
\title{Show the NEWS and documentation files for the SuperLearner package}
\usage{
SuperLearnerDocs(what = "SuperLearnerR.pdf", ...)

SuperLearnerNews(...)
}
\arguments{
\item{what}{specify what document to open. Currently supports the NEWS file
and the PDF files 'SuperLearner.pdf' and 'SuperLearnerR.pdf'.}

\item{\dots}{additional arguments passed to \code{\link[utils:RShowDoc]{utils::RShowDoc()}}.}
}
\value{
A invisible character string given the path to the SuperLearner NEWS or documentation file.
}
\description{
Show the NEWS file of the SuperLearner package. The function is simply a
wrapper for \code{\link[utils:RShowDoc]{utils::RShowDoc()}}.
}
\keyword{utilities}
