% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.cforest.R
\name{SL.cforest}
\alias{SL.cforest}
\alias{predict.SL.cforest}
\title{SL wrapper for \code{party::cforest()}}
\usage{
SL.cforest(
  Y,
  X,
  newX = X,
  family = gaussian(),
  obsWeights = NULL,
  ntree = 1000,
  mtry = max(floor(ncol(X)/3), 1),
  mincriterion = 0,
  teststat = "quad",
  testtype = "Univ",
  replace = F,
  fraction = 0.632,
  ...
)

\method{predict}{SL.cforest}(object, newdata, ...)
}
\arguments{
\item{Y}{The outcome in the training data set. Must be a numeric vector.}

\item{X}{The predictor variables in the training data set, usually a data.frame.}

\item{newX}{The predictor variables in the validation data set. The
structure should match X.}

\item{family}{Either \code{\link[=gaussian]{gaussian()}} or \code{\link[=binomial]{binomial()}} to
describe the error distribution. Link function information will be ignored.}

\item{obsWeights}{Optional observation weights.}

\item{ntree}{Number of trees.}

\item{mtry}{Number of randomly selected features per node.}

\item{mincriterion, teststat, testtype, replace, fraction}{Arguments passed to \code{party::cforest_control()}.}

\item{...}{For SL wrappers, other remaining arguments. For \code{write.SL.template()}, arguments passed to \code{\link[=cat]{cat()}}.}

\item{object}{an object resulting from the fitting function.}

\item{newdata}{a data frame of predictors for which to compute predictions.}
}
\description{
These defaults emulate \code{party::cforest_unbiased()} but allow customization.
}
