% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SL.speedglm.R
\name{SL.speedglm}
\alias{SL.speedglm}
\title{Wrapper for speedglm}
\usage{
SL.speedglm(Y, X, newX, family, obsWeights, maxit = 25, k = 2, ...)
}
\arguments{
\item{Y}{Outcome variable}

\item{X}{Training dataframe}

\item{newX}{Test dataframe}

\item{family}{Gaussian or binomial}

\item{obsWeights}{Observation-level weights}

\item{maxit}{Maximum number of iterations before stopping.}

\item{k}{numeric, the penalty per parameter to be used; the default k = 2 is
the classical AIC.}

\item{...}{Any remaining arguments, not used.}
}
\description{
Speedglm is a fast version of glm()
}
\references{
Enea, M. A. R. C. O. (2013). Fitting linear models and generalized linear
models with large data sets in R. Statistical Methods for the Analysis of
Large Datasets: book of short papers, 411-414.
}
\seealso{
\code{\link{predict.SL.speedglm}} \code{\link[speedglm]{speedglm}}
  \code{\link[speedglm]{predict.speedglm}}
}
